Change unit tests: * extract CloudItemMetadataProvider to its own class
