+++ src/test/java/org/cryptomator/fusecloudaccess/AttributesTest.java
@@ -1,7 +1,6 @@
 package org.cryptomator.fusecloudaccess;
 
 import jnr.ffi.Runtime;
-import org.cryptomator.cloudaccess.api.CloudItemMetadata;
 import org.cryptomator.cloudaccess.api.CloudItemType;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -12,10 +11,8 @@ import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import ru.serce.jnrfuse.struct.FileStat;
 
-import java.nio.file.Path;
 import java.time.Instant;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Stream;
 
 
@@ -70,40 +67,6 @@ public class AttributesTest {
 		Attributes.copy(CloudItemMetadataProvider.ofSize(size), fileStat);
 		Assertions.assertEquals(expectedSize, fileStat.st_size.longValue());
 
-	}
-
-
-	private static class CloudItemMetadataProvider {
-
-		private static final String DEFAULT_NAME = "unknown.obj";
-		private static final Path DEFAULT_PATH = Path.of("path/to/object");
-		private static final CloudItemType DEFAULT_TYPE = CloudItemType.UNKNOWN;
-		private static final Instant DEFAULT_MTIME = Instant.now();
-		private static final long DEFAULT_SIZE = 1337L;
-
-		public static CloudItemMetadata of(String name, Path p, CloudItemType type, Instant mTime, Long size) {
-			return new CloudItemMetadata(name, p, type, Optional.ofNullable(mTime), Optional.ofNullable(size));
-		}
-
-		public static CloudItemMetadata ofName(String name) {
-			return CloudItemMetadataProvider.of(name, DEFAULT_PATH, DEFAULT_TYPE, DEFAULT_MTIME, DEFAULT_SIZE);
-		}
-
-		public static CloudItemMetadata ofPath(Path p) {
-			return CloudItemMetadataProvider.of(DEFAULT_NAME, p, DEFAULT_TYPE, DEFAULT_MTIME, DEFAULT_SIZE);
-		}
-
-		public static CloudItemMetadata ofType(CloudItemType type) {
-			return CloudItemMetadataProvider.of(DEFAULT_NAME, DEFAULT_PATH, type, DEFAULT_MTIME, DEFAULT_SIZE);
-		}
-
-		public static CloudItemMetadata ofMTime(Instant mTime) {
-			return CloudItemMetadataProvider.of(DEFAULT_NAME, DEFAULT_PATH, DEFAULT_TYPE, mTime, DEFAULT_SIZE);
-		}
-
-		public static CloudItemMetadata ofSize(Long size) {
-			return CloudItemMetadataProvider.of(DEFAULT_NAME, DEFAULT_PATH, DEFAULT_TYPE, DEFAULT_MTIME, size);
-		}
 	}
 
 }
